Refactor: separate time utils and variables
